Extract usePrettifyQuery hook
